(not checked) Fixed race condition in voting
before: Mafia will vote in daytime, and then wait until everyone else is finished voting and then continue. However, the voting finished, went to nighttime, and since one mafia went straight to voting. The mafia loop didn't catch this voting switch, and is_time_to_vote() considers nightime_voting and daytime_voting to be the same, so then it didn't catch the change in voting, and so kept waiting, holding up everything.

Now: Fixed so that it will check that the voting phase didn't change, and is properly waiting for the same phase_voting to end.